Change all usage references to use default_version
Following the discussion on https://github.com/denoland/deno_website2/pull/1132
I'm not sure this will be merged soon, so maybe it makes sense
to update the documentation for now?

Adding a version to the import is imho not a bad behaviour to
begin with, even if it to explicit specifing `main` as default
version.